Re-factor health.go. Closes #2588
