+++ controller/sync_hooks_test.go
@@ -1,0 +1,45 @@
+package controller
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+
+	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
+)
+
+func TestGenericPhase(t *testing.T) {
+	sampleWorkflow := unstructured.Unstructured{Object: map[string]interface{}{
+		"spec": map[string]interface{}{
+			"entrypoint":    "sampleEntryPoint",
+			"extraneousKey": "we are agnostic to extraneous keys",
+		},
+		"status": map[string]interface{}{
+			"phase":   "Running",
+			"message": "This node is running",
+		},
+	},
+	}
+
+	status, message := getGenericPhase(&sampleWorkflow)
+	assert.Equal(t, v1alpha1.OperationRunning, status)
+	assert.Equal(t, "This node is running", message)
+
+	sampleWorkflow = unstructured.Unstructured{Object: map[string]interface{}{
+		"spec": map[string]interface{}{
+			"entrypoint":    "sampleEntryPoint",
+			"extraneousKey": "we are agnostic to extraneous keys",
+		},
+		"status": map[string]interface{}{
+			"phase":   "Succeeded",
+			"message": "This node is has succeeded",
+		},
+	},
+	}
+
+	status, message = getGenericPhase(&sampleWorkflow)
+	assert.Equal(t, v1alpha1.OperationSucceeded, status)
+	assert.Equal(t, "This node is has succeeded", message)
+
+}
