+++ util/health/daemonset_test.go
@@ -1,0 +1,9 @@
+package health
+
+import (
+	"testing"
+)
+
+func Test_getDaemonSetHealth(t *testing.T) {
+	// TODO
+}
