+++ util/health/replicaset_test.go
@@ -1,0 +1,9 @@
+package health
+
+import (
+	"testing"
+)
+
+func Test_getReplicaSetHealth(t *testing.T) {
+	// TODO
+}
