+++ util/health/statefulset_test.go
@@ -1,0 +1,11 @@
+package health
+
+import (
+	"testing"
+
+	appv1 "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
+)
+
+func TestStatefulSetHealth(t *testing.T) {
+	assertAppHealth(t, "./testdata/statefulset.yaml", appv1.HealthStatusHealthy)
+}
